check-pr-commits.sh: check for empty commit body
Problem: The check-pr-commits.sh script does not check for an empty
commit body, but only in very rare circumstances should a commit be
allowed without a body.

Fail the pr-validator check if an empty commit message body is
detected.  This will encourage contributors to explain each commit,
even if briefly.