check-pr-commits.sh: allow subtree merge commits by special case
Problem: The PR validator rejects PRs that have incorporated third
party repos with `git subtree` due to the use of subtree merge commits,
which are merge commits, have no body, and the subject will typically
exceed the hard limit of 70 characters.

Add a new function for detecting subtree merge commits by looking for
the pattern `Merge commit .* as .*`, which appears to be specific to
subtree merges. Then, use this function to special case all the problem
areas above so that subtree merges (and hopefully only subtree merges)
are allowed.