move workaround for git dubious ownership to top of script
Problem: On projects with a `main` branch instead of `master`
`check-pr-commits.sh` fails because `origin_main()` always returns
`origin/master` due to a "dubious ownership" from git.  The workaround
for this issue is run too late in the script, after `$BASE` has
already been computed.

Move the fix for the "dubious ownership" error to the top of the
script so that it runs, if necessary, before the first git command.